kubernetes-node/micro-services/backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
drf-yasg==1.21.7